Pilot-Templates/wx-orchestrate/external/code-engine/requirements.txt
fastmcp
pillow
python-dotenv